✨ feat: add sticker maker
This commit adds a new feature to the WhatsApp bot that allows users to create stickers from images.

The new feature is implemented using the `wa-sticker-formatter` library, which provides a simple and easy-to-use API for creating stickers.

To use the new feature, users can simply send an image to the bot with the command `.s`. The bot will then automatically create a sticker from the image and send it back to the user.

This new feature is a great way for users to add some fun and personality to their WhatsApp conversations.