feat(whatsapp-action): implement "hello world" action
This commit introduces a new "hello world" action to the WhatsApp bot.
The action is implemented using the `@WhatsappMessage()` decorator and can be triggered by sending a message with the text "hello" to the bot.
When triggered, the action will send a message back to the user saying "hello world".