feat(whatsapp): introduces IsEligible decorator
This commit introduces the @IsEligible decorator to the Whatsapp module. This decorator allows developers to specify which conditions should be met for a message to be processed by a WhatsappMessageAction.

The IsEligible decorator can be used in combination with the existing @WhatsappMessageAction decorator. To use the IsEligible decorator, simply add it to the method that you want to be executed only if the specified conditions are met.

For example:

```typescript
@WhatsappMessage({
 flags: 'hello',
})
export class HelloWorldAction extends WhatsappMessageAction {
 @IsEligible()
 public onlyMe(socket: WASocket, message: WAMessage) {
  console.log('only me');
  console.log(message);
  console.log('only me');
  return true;
 }
}
```

In this example, the onlyMe method will only be executed if the message is from the same person who sent the message that triggered the WhatsappMessageAction.